Agent_Movement_Improved
Agent can set Recently Visited attribute of 2 Cell before the current to false properly. Stress level added. So that if Agent stuck on 2 Cell because there is no safe and more valuable Cell from that 2 Cell, Agent breaks the it's chains and takes action to choosing dangerous but not deadly Cell instead of that 2 Cell.